refactor: model api dependency correctly
