+++ wip/Sources/API/APIClientLive.swift
@@ -6,13 +6,28 @@
 //
 
 import Foundation
+import Alamofire
+import WIPKit
 
 extension APIClient {
     public static var liveValue: Self {
-        let apiManager = APIManager()
+        let baseURL = "http://127.0.0.1:8080/"
+
         return Self(
             addNewProject: { project in
-                return try await apiManager.postProject(project)
+                let request = AF.request(baseURL+APIEndpoints.projects, method: .post, parameters: project, encoder: JSONParameterEncoder.prettyPrinted)
+
+                let dataTask = request.serializingDecodable( Project.self)
+                let response = await dataTask.result
+                do {
+                    return try response.get()
+                } catch {
+                    print(error)
+                    throw APIError()
+                }
+            },
+            fetchAllProjects: {
+                return [Project]()
             }
         )
     }
